remove pattern for document selector in LanguageClientOptions (#1137)
This pr removes the filter pattern for document selector in LanguageClientOptions. It will then filter out all undefined results out of the responses provided the the different LS's.

middleware and more particulary middlewareProvideDefinition were added to remove the unwanted responses returned by all the started LS:
As all the language servers now react to the textDocument_definition call, a lot of them don't have the asked definition in their anchormap. This causes them to return "/undefined". Without the middleware, when trying to navigate with ctrl+click, vscode would propose (with a popup) all the location returned by the different LSes. which resulted in the right location and a "/undefined" option. The middlewareProvideDefinition will remove all the "/undefined" propositions.

closes #1136